Core/Spells: Fixed spell effect value truncation when value is a large integer
Port From (https://github.com/TrinityCore/TrinityCore/commit/2d205506bfd3a2bb253db73b121fd147928811ba)

+++ Source/Game/Entities/Object/WorldObject.cs
@@ -1827,7 +1827,7 @@ namespace Game.Entities
             return spellInfo.GetMinRange(!IsHostileTo(target));
         }
 
-        public float ApplyEffectModifiers(SpellInfo spellInfo, uint effIndex, float value)
+        public double ApplyEffectModifiers(SpellInfo spellInfo, uint effIndex, double value)
         {
             Player modOwner = GetSpellModOwner();
             if (modOwner != null)
